refactor(api): 아키텍처 일관성 정리 — 리뷰 후속 4건
아키텍처 일관성 리뷰에서 나온 후속 정리:
- withTimeout을 common/redis → common/utils/with-timeout.util.ts로 이동
  (Redis 무관 범용 유틸 — mask.util 선례)
- toErrorMessage 유틸 신설(common/utils/error-message.util.ts), 샘플러·
  jwt.strategy·bull.health·non-blocking-init의 중복 toMessage 4곳 통합
- 로그 샘플러 메시지에서 '(fail-open)' 하드코딩 제거 — fail-closed 계약인
  락 어댑터도 같은 샘플러를 쓰므로 정책 문구는 중립 유지
- e2e-app-factory가 REDIS_COMMAND_CLIENT도 RedisMock으로 오버라이드
  (E2E 앱이 localhost:6379로 실연결 시도하던 잔여 핸들 제거)
