+++ apps/api/src/common/bullmq/non-blocking-init.ts
@@ -1,4 +1,5 @@
 import type { Logger } from "@nestjs/common";
+import { toErrorMessage } from "../utils/error-message.util";
 
 /**
  * 부팅을 블로킹하지 않는 초기화 실행 (fire-and-forget)
@@ -20,8 +21,6 @@ export async function runInBackground(
 		// async 함수라 task()의 동기 throw도 catch로 흡수된다 (부트스트랩 크래시 방지)
 		await task();
 	} catch (error: unknown) {
-		logger.error(
-			`${label} failed: ${error instanceof Error ? error.message : String(error)}`,
-		);
+		logger.error(`${label} failed: ${toErrorMessage(error)}`);
 	}
 }
